src/modules/rentals/useCase/createRental/CreateRentalUseCase.ts
import { ICarsRepository } from "@modules/cars/repositories/ICarsRepository";
import { Rental } from "@modules/rentals/infra/typeorm/entities/Rental";
import { IRentalRepository } from "@modules/rentals/repositories/IRentalRepository";
import { IDateProvider } from "@shared/container/provider/IDateProvider";
import { AppError } from "@shared/error/AppError";
import { inject, injectable } from "tsyringe";

interface IRequest {
  user_id: string;
  car_id: string;
  expected_return_date: Date;
}

@injectable()
class CreateRentalUseCase {
  constructor(
    @inject("RentalRepository")
    private rentalsRepository: IRentalRepository,
    @inject("DayjsDateProvider")
    private dayDateProvider: IDateProvider,
    @inject("CarsRepository")
    private carsRepositiory: ICarsRepository
  ) {}

  async execute({
    car_id,
    expected_return_date,
    user_id,
  }: IRequest): Promise<Rental> {
    const minimumHour = 24;

    const carUnavaliable = await this.rentalsRepository.findOpenRentalByCar(
      car_id
    );

    if (carUnavaliable) {
      throw new AppError("Car is unavaliable");
    }

    const rentalOpenToUser = await this.rentalsRepository.findOpenRentalByUser(
      user_id
    );

    if (rentalOpenToUser) {
      throw new AppError("There's a rental in progress for user!");
    }

    const dateNow = this.dayDateProvider.dateNow();

    const compare = this.dayDateProvider.compareInHours(
      dateNow,
      expected_return_date
    );

    if (compare < minimumHour) {
      throw new AppError("Invalid return time!");
    }

    //console.log("Compare Date", compare);

    const rental = await this.rentalsRepository.create({
      user_id,
      car_id,
      expected_return_date,
    });

    await this.carsRepositiory.updateAvailable(car_id, false);

    return rental;
  }
}

export { CreateRentalUseCase };
